fix: height is not accurate when revealing
Sometimes reveal widgets can be hidden, for example it can be in the tabbed component which has some hidden tabs or you can say their CSS display is none, so if this reveal widget is on those tabs then it will be hidden and the height will not be accurate when the JS is initialize it will be 0. So rather then saving the original height into a variable we can find out it's real height on the fly.